requirements.txt
flask>=3.0.0
Pillow
flask-limiter
gunicorn
